+++ bot/src/commonMain/kotlin/me/leondorus/data/UserId.kt
@@ -1,7 +1,7 @@
 package me.leondorus.data
 
 @JvmInline
-value class UserId(val id: Int)
+value class UserId(val id: Long)
 
-val Int.uid: UserId
+val Long.uid: UserId
     get() = UserId(this)
