+++ bot/src/commonMain/kotlin/me/leondorus/room/react/ReactDao.kt
@@ -15,5 +15,5 @@ interface ReactDao {
     suspend fun delete(react: RoomReact)
 
     @Query("SELECT * FROM reacts WHERE messageId = :mid")
-    fun getByMessageId(mid: Int): Flow<List<RoomReact>>
+    fun getByMessageId(mid: Long): Flow<List<RoomReact>>
 }
